refactor: extract SubsonicPlaylistController — fixes #104
