+++ airsonic-main/src/main/java/org/airsonic/player/controller/SubsonicPlaylistController.java
@@ -1,0 +1,335 @@
+/*
+ This file is part of Airsonic.
+
+ Airsonic is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ Airsonic is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with Airsonic.  If not, see <http://www.gnu.org/licenses/>.
+
+ Copyright 2025 (C) Airsonic Authors
+ Based upon Subsonic, Copyright 2009 (C) Sindre Mehus
+ */
+package org.airsonic.player.controller;
+
+import org.airsonic.player.domain.MediaFile;
+import org.airsonic.player.domain.Player;
+import org.airsonic.player.domain.PlayerTechnology;
+import org.airsonic.player.service.JaxbContentService;
+import org.airsonic.player.service.MediaFileService;
+import org.airsonic.player.service.PlayerService;
+import org.airsonic.player.service.PlaylistService;
+import org.airsonic.player.service.SecurityService;
+import org.airsonic.player.util.StringUtil;
+import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.subsonic.restapi.PlaylistWithSongs;
+import org.subsonic.restapi.Playlists;
+import org.subsonic.restapi.Response;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequestWrapper;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.SortedSet;
+import java.util.TreeSet;
+
+import static org.springframework.web.bind.ServletRequestUtils.getBooleanParameter;
+import static org.springframework.web.bind.ServletRequestUtils.getIntParameter;
+import static org.springframework.web.bind.ServletRequestUtils.getIntParameters;
+import static org.springframework.web.bind.ServletRequestUtils.getRequiredIntParameter;
+
+@Controller
+@RequestMapping(value = {"/rest", "/ext"}, method = {RequestMethod.GET, RequestMethod.POST})
+public class SubsonicPlaylistController {
+
+    @Autowired
+    private JAXBWriter jaxbWriter;
+    @Autowired
+    private PlaylistService playlistService;
+    @Autowired
+    private MediaFileService mediaFileService;
+    @Autowired
+    private PlayerService playerService;
+    @Autowired
+    private SecurityService securityService;
+    @Autowired
+    private JaxbContentService jaxbContentService;
+
+    @ExceptionHandler(MissingServletRequestParameterException.class)
+    public void handleMissingRequestParam(HttpServletRequest request,
+                                          HttpServletResponse response,
+                                          MissingServletRequestParameterException exception) {
+        error(request, response, SubsonicRESTController.ErrorCode.MISSING_PARAMETER,
+                "Required param (" + exception.getParameterName() + ") is missing");
+    }
+
+    @RequestMapping({"/getPlaylists", "/getPlaylists.view"})
+    public void getPlaylists(HttpServletRequest request, HttpServletResponse response) {
+        request = wrapRequest(request);
+
+        org.airsonic.player.domain.User user = securityService.getCurrentUser(request);
+        String authenticatedUsername = user.getUsername();
+        String requestedUsername = request.getParameter("username");
+
+        if (requestedUsername == null) {
+            requestedUsername = authenticatedUsername;
+        } else if (!user.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, authenticatedUsername + " is not authorized to get playlists for " + requestedUsername);
+            return;
+        }
+
+        Playlists result = new Playlists();
+
+        for (org.airsonic.player.domain.Playlist playlist : playlistService.getReadablePlaylistsForUser(requestedUsername)) {
+            result.getPlaylist().add(jaxbContentService.createJaxbPlaylist(new org.subsonic.restapi.Playlist(), playlist));
+        }
+
+        Response res = createResponse();
+        res.setPlaylists(result);
+        jaxbWriter.writeResponse(request, response, res);
+    }
+
+    @RequestMapping({"/getPlaylist", "/getPlaylist.view"})
+    public void getPlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        request = wrapRequest(request);
+        String username = securityService.getCurrentUsername(request);
+        Player player = playerService.getPlayer(request, response, username);
+
+        int id = getRequiredIntParameter(request, "id");
+
+        org.airsonic.player.domain.Playlist playlist = playlistService.getPlaylist(id);
+        if (playlist == null) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "Playlist not found: " + id);
+            return;
+        }
+        if (!playlistService.isReadAllowed(playlist, username)) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
+            return;
+        }
+        PlaylistWithSongs result = jaxbContentService.createJaxbPlaylist(new PlaylistWithSongs(), playlist);
+        for (MediaFile mediaFile : playlistService.getFilesInPlaylist(id)) {
+            if (securityService.isFolderAccessAllowed(mediaFile, username)) {
+                result.getEntry().add(jaxbContentService.createJaxbChild(player, mediaFile, username));
+            }
+        }
+
+        Response res = createResponse();
+        res.setPlaylist(result);
+        jaxbWriter.writeResponse(request, response, res);
+    }
+
+    @RequestMapping({"/createPlaylist", "/createPlaylist.view"})
+    public void createPlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        request = wrapRequest(request, true);
+        String username = securityService.getCurrentUsername(request);
+
+        Integer playlistId = getIntParameter(request, "playlistId");
+        String name = request.getParameter("name");
+        if (playlistId == null && name == null) {
+            error(request, response, SubsonicRESTController.ErrorCode.MISSING_PARAMETER, "Playlist ID or name must be specified.");
+            return;
+        }
+
+        org.airsonic.player.domain.Playlist playlist;
+        if (playlistId != null) {
+            playlist = playlistService.getPlaylist(playlistId);
+            if (playlist == null) {
+                error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "Playlist not found: " + playlistId);
+                return;
+            }
+            if (!playlistService.isWriteAllowed(playlist, username)) {
+                error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + playlistId);
+                return;
+            }
+        } else {
+            playlist = playlistService.createPlaylist(name, false, username);
+            playlistService.broadcast(playlist);
+        }
+
+        List<MediaFile> songs = new ArrayList<MediaFile>();
+        for (int id : getIntParameters(request, "songId")) {
+            MediaFile song = mediaFileService.getMediaFile(id);
+            if (song != null) {
+                songs.add(song);
+            }
+        }
+        playlistService.setFilesInPlaylist(playlist.getId(), songs);
+        playlistService.broadcastFileChange(playlist.getId(), false, true);
+
+        writeEmptyResponse(request, response);
+    }
+
+    @RequestMapping({"/updatePlaylist", "/updatePlaylist.view"})
+    public void updatePlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        request = wrapRequest(request, true);
+        String username = securityService.getCurrentUsername(request);
+
+        int id = getRequiredIntParameter(request, "playlistId");
+        org.airsonic.player.domain.Playlist playlist = playlistService.getPlaylist(id);
+        if (playlist == null) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "Playlist not found: " + id);
+            return;
+        }
+        if (!playlistService.isWriteAllowed(playlist, username)) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
+            return;
+        }
+
+        // create new object to not mutate the cache
+        String name = request.getParameter("name");
+        String comment = request.getParameter("comment");
+        Boolean shared = getBooleanParameter(request, "public");
+        playlistService.updatePlaylist(id, name, comment, shared);
+
+        // TODO: Add later
+//            for (String usernameToAdd : ServletRequestUtils.getStringParameters(request, "usernameToAdd")) {
+//                if (securityService.getUserByName(usernameToAdd) != null) {
+//                    playlistService.addPlaylistUser(id, usernameToAdd);
+//                }
+//            }
+//            for (String usernameToRemove : ServletRequestUtils.getStringParameters(request, "usernameToRemove")) {
+//                if (securityService.getUserByName(usernameToRemove) != null) {
+//                    playlistService.deletePlaylistUser(id, usernameToRemove);
+//                }
+//            }
+        List<MediaFile> songs = playlistService.getFilesInPlaylist(id);
+        boolean songsChanged = false;
+
+        SortedSet<Integer> tmp = new TreeSet<Integer>();
+        for (int songIndexToRemove : getIntParameters(request, "songIndexToRemove")) {
+            tmp.add(songIndexToRemove);
+        }
+        List<Integer> songIndexesToRemove = new ArrayList<Integer>(tmp);
+        Collections.reverse(songIndexesToRemove);
+        for (Integer songIndexToRemove : songIndexesToRemove) {
+            songs.remove(songIndexToRemove.intValue());
+            songsChanged = true;
+        }
+        for (int songToAdd : getIntParameters(request, "songIdToAdd")) {
+            MediaFile song = mediaFileService.getMediaFile(songToAdd);
+            if (song != null) {
+                songs.add(song);
+                songsChanged = true;
+            }
+        }
+        if (songsChanged) {
+            playlistService.setFilesInPlaylist(id, songs);
+        }
+        playlistService.broadcastFileChange(id, playlist.getShared(), songsChanged);
+
+        writeEmptyResponse(request, response);
+    }
+
+    @RequestMapping({"/deletePlaylist", "/deletePlaylist.view"})
+    public void deletePlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        request = wrapRequest(request, true);
+        String username = securityService.getCurrentUsername(request);
+
+        int id = getRequiredIntParameter(request, "id");
+        if (!playlistService.isExist(id)) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "Playlist not found: " + id);
+            return;
+        }
+        if (!playlistService.isWriteAllowed(id, username)) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
+            return;
+        }
+        playlistService.deletePlaylist(id);
+        playlistService.broadcastDeleted(id);
+
+        writeEmptyResponse(request, response);
+    }
+
+    private Response createResponse() {
+        return jaxbWriter.createResponse(true);
+    }
+
+    private void writeEmptyResponse(HttpServletRequest request, HttpServletResponse response) {
+        jaxbWriter.writeResponse(request, response, createResponse());
+    }
+
+    private void error(HttpServletRequest request, HttpServletResponse response,
+                       SubsonicRESTController.ErrorCode code, String message) {
+        jaxbWriter.writeErrorResponse(request, response, code, message);
+    }
+
+    private HttpServletRequest wrapRequest(HttpServletRequest request) {
+        return wrapRequest(request, false);
+    }
+
+    private HttpServletRequest wrapRequest(final HttpServletRequest request, boolean jukebox) {
+        final Integer playerId = createPlayerIfNecessary(request, jukebox);
+        return new HttpServletRequestWrapper(request) {
+            @Override
+            public String getParameter(String name) {
+                // Returns the correct player to be used in PlayerService.getPlayer()
+                if ("player".equals(name)) {
+                    return playerId == null ? null : String.valueOf(playerId);
+                }
+
+                // Support old style ID parameters.
+                if ("id".equals(name)) {
+                    return mapId(request.getParameter("id"));
+                }
+
+                return super.getParameter(name);
+            }
+        };
+    }
+
+    private String mapId(String id) {
+        if (id == null || id.startsWith(CoverArtController.ALBUM_COVERART_PREFIX) ||
+                id.startsWith(CoverArtController.ARTIST_COVERART_PREFIX) || StringUtils.isNumeric(id)) {
+            return id;
+        }
+
+        try {
+            String path = StringUtil.utf8HexDecode(id);
+            MediaFile mediaFile = mediaFileService.getMediaFile(path);
+            return String.valueOf(mediaFile.getId());
+        } catch (Exception x) {
+            return id;
+        }
+    }
+
+    private Integer createPlayerIfNecessary(HttpServletRequest request, boolean jukebox) {
+        String username = request.getRemoteUser();
+        String clientId = request.getParameter("c");
+        if (jukebox) {
+            clientId += "-jukebox";
+        }
+
+        List<Player> players = playerService.getPlayersForUserAndClientId(username, clientId);
+
+        // If not found, create it.
+        if (players.isEmpty()) {
+            Player player = new Player();
+            player.setIpAddress(request.getRemoteAddr());
+            player.setUsername(username);
+            player.setClientId(clientId);
+            player.setName(clientId);
+            player.setTechnology(jukebox ? PlayerTechnology.JUKEBOX : PlayerTechnology.EXTERNAL_WITH_PLAYLIST);
+            playerService.createPlayer(player);
+            players = playerService.getPlayersForUserAndClientId(username, clientId);
+        }
+
+        // Return the player ID.
+        return !players.isEmpty() ? players.get(0).getId() : null;
+    }
+}
