+++ airsonic-main/src/main/java/org/airsonic/player/controller/SubsonicRESTController.java
@@ -94,8 +94,6 @@ public class SubsonicRESTController {
     @Autowired
     private HLSController hlsController;
     @Autowired
-    private PlaylistService playlistService;
-    @Autowired
     private LyricsService lyricsService;
     @Autowired
     private ArtistService artistService;
@@ -304,183 +302,6 @@ public class SubsonicRESTController {
 
 
 
-
-    @RequestMapping({"/getPlaylists", "/getPlaylists.view"})
-    public void getPlaylists(HttpServletRequest request, HttpServletResponse response) {
-        request = wrapRequest(request);
-
-        org.airsonic.player.domain.User user = securityService.getCurrentUser(request);
-        String authenticatedUsername = user.getUsername();
-        String requestedUsername = request.getParameter("username");
-
-        if (requestedUsername == null) {
-            requestedUsername = authenticatedUsername;
-        } else if (!user.isAdminRole()) {
-            error(request, response, ErrorCode.NOT_AUTHORIZED, authenticatedUsername + " is not authorized to get playlists for " + requestedUsername);
-            return;
-        }
-
-        Playlists result = new Playlists();
-
-        for (org.airsonic.player.domain.Playlist playlist : playlistService.getReadablePlaylistsForUser(requestedUsername)) {
-            result.getPlaylist().add(jaxbContentService.createJaxbPlaylist(new org.subsonic.restapi.Playlist(), playlist));
-        }
-
-        Response res = createResponse();
-        res.setPlaylists(result);
-        jaxbWriter.writeResponse(request, response, res);
-    }
-
-    @RequestMapping({"/getPlaylist", "/getPlaylist.view"})
-    public void getPlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request);
-        String username = securityService.getCurrentUsername(request);
-        Player player = playerService.getPlayer(request, response, username);
-
-        int id = getRequiredIntParameter(request, "id");
-
-        org.airsonic.player.domain.Playlist playlist = playlistService.getPlaylist(id);
-        if (playlist == null) {
-            error(request, response, ErrorCode.NOT_FOUND, "Playlist not found: " + id);
-            return;
-        }
-        if (!playlistService.isReadAllowed(playlist, username)) {
-            error(request, response, ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
-            return;
-        }
-        PlaylistWithSongs result = jaxbContentService.createJaxbPlaylist(new PlaylistWithSongs(), playlist);
-        for (MediaFile mediaFile : playlistService.getFilesInPlaylist(id)) {
-            if (securityService.isFolderAccessAllowed(mediaFile, username)) {
-                result.getEntry().add(jaxbContentService.createJaxbChild(player, mediaFile, username));
-            }
-        }
-
-        Response res = createResponse();
-        res.setPlaylist(result);
-        jaxbWriter.writeResponse(request, response, res);
-    }
-
-    @RequestMapping({"/createPlaylist", "/createPlaylist.view"})
-    public void createPlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request, true);
-        String username = securityService.getCurrentUsername(request);
-
-        Integer playlistId = getIntParameter(request, "playlistId");
-        String name = request.getParameter("name");
-        if (playlistId == null && name == null) {
-            error(request, response, ErrorCode.MISSING_PARAMETER, "Playlist ID or name must be specified.");
-            return;
-        }
-
-        org.airsonic.player.domain.Playlist playlist;
-        if (playlistId != null) {
-            playlist = playlistService.getPlaylist(playlistId);
-            if (playlist == null) {
-                error(request, response, ErrorCode.NOT_FOUND, "Playlist not found: " + playlistId);
-                return;
-            }
-            if (!playlistService.isWriteAllowed(playlist, username)) {
-                error(request, response, ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + playlistId);
-                return;
-            }
-        } else {
-            playlist = playlistService.createPlaylist(name, false, username);
-            playlistService.broadcast(playlist);
-        }
-
-        List<MediaFile> songs = new ArrayList<MediaFile>();
-        for (int id : getIntParameters(request, "songId")) {
-            MediaFile song = mediaFileService.getMediaFile(id);
-            if (song != null) {
-                songs.add(song);
-            }
-        }
-        playlistService.setFilesInPlaylist(playlist.getId(), songs);
-        playlistService.broadcastFileChange(playlist.getId(), false, true);
-
-        writeEmptyResponse(request, response);
-    }
-
-    @RequestMapping({"/updatePlaylist", "/updatePlaylist.view"})
-    public void updatePlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request, true);
-        String username = securityService.getCurrentUsername(request);
-
-        int id = getRequiredIntParameter(request, "playlistId");
-        org.airsonic.player.domain.Playlist playlist = playlistService.getPlaylist(id);
-        if (playlist == null) {
-            error(request, response, ErrorCode.NOT_FOUND, "Playlist not found: " + id);
-            return;
-        }
-        if (!playlistService.isWriteAllowed(playlist, username)) {
-            error(request, response, ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
-            return;
-        }
-
-        // create new object to not mutate the cache
-        String name = request.getParameter("name");
-        String comment = request.getParameter("comment");
-        Boolean shared = getBooleanParameter(request, "public");
-        playlistService.updatePlaylist(id, name, comment, shared);
-
-        // TODO: Add later
-//            for (String usernameToAdd : ServletRequestUtils.getStringParameters(request, "usernameToAdd")) {
-//                if (securityService.getUserByName(usernameToAdd) != null) {
-//                    playlistService.addPlaylistUser(id, usernameToAdd);
-//                }
-//            }
-//            for (String usernameToRemove : ServletRequestUtils.getStringParameters(request, "usernameToRemove")) {
-//                if (securityService.getUserByName(usernameToRemove) != null) {
-//                    playlistService.deletePlaylistUser(id, usernameToRemove);
-//                }
-//            }
-        List<MediaFile> songs = playlistService.getFilesInPlaylist(id);
-        boolean songsChanged = false;
-
-        SortedSet<Integer> tmp = new TreeSet<Integer>();
-        for (int songIndexToRemove : getIntParameters(request, "songIndexToRemove")) {
-            tmp.add(songIndexToRemove);
-        }
-        List<Integer> songIndexesToRemove = new ArrayList<Integer>(tmp);
-        Collections.reverse(songIndexesToRemove);
-        for (Integer songIndexToRemove : songIndexesToRemove) {
-            songs.remove(songIndexToRemove.intValue());
-            songsChanged = true;
-        }
-        for (int songToAdd : getIntParameters(request, "songIdToAdd")) {
-            MediaFile song = mediaFileService.getMediaFile(songToAdd);
-            if (song != null) {
-                songs.add(song);
-                songsChanged = true;
-            }
-        }
-        if (songsChanged) {
-            playlistService.setFilesInPlaylist(id, songs);
-        }
-        playlistService.broadcastFileChange(id, playlist.getShared(), songsChanged);
-
-        writeEmptyResponse(request, response);
-    }
-
-    @RequestMapping({"/deletePlaylist", "/deletePlaylist.view"})
-    public void deletePlaylist(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        request = wrapRequest(request, true);
-        String username = securityService.getCurrentUsername(request);
-
-        int id = getRequiredIntParameter(request, "id");
-        if (!playlistService.isExist(id)) {
-            error(request, response, ErrorCode.NOT_FOUND, "Playlist not found: " + id);
-            return;
-        }
-        if (!playlistService.isWriteAllowed(id, username)) {
-            error(request, response, ErrorCode.NOT_AUTHORIZED, "Permission denied for playlist " + id);
-            return;
-        }
-        playlistService.deletePlaylist(id);
-        playlistService.broadcastDeleted(id);
-
-        writeEmptyResponse(request, response);
-    }
 
     @RequestMapping({"/download", "/download.view"})
     public ResponseEntity<Resource> download(Principal p,
